refactor: 청원 삭제 권한 설정 (#58)
