+++ src/main/java/com/smunity/api/domain/petition/service/PetitionService.java
@@ -9,5 +9,5 @@ public interface PetitionService {
     PetitionDto getPetition(Long id);
     PetitionDto createPetition(PetitionDto petitionDto, String token);
     PetitionDto changePetition(Long id, PetitionDto petitionDto, String token);
-    void deletePetition(Long id);
+    void deletePetition(Long id, String token);
 }
